requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
selenium>=4.15.2
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
requests>=2.31.0
webdriver-manager>=4.0.1
pydantic-settings>=2.0.3
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
jinja2>=3.1.2
aiofiles>=23.2.1
python-multipart>=0.0.6
APScheduler==3.10.4
openpyxl==3.1.2
pandas==2.1.4
tenacity>=8.2.3
pytz>=2023.3
google-api-python-client>=2.108.0
google-auth>=2.23.4
google-auth-httplib2>=0.2.0
importlib-metadata>=6.8.0
